added DTO dependencies for users & addUser method working with DB
