+++ src/main/java/com/example/gwent_projet/GwentProjetApplication.java
@@ -1,5 +1,6 @@
 package com.example.gwent_projet;
 
+import com.example.gwent_projet.cli.UserCommand;
 import com.example.gwent_projet.models.Card;
 import com.example.gwent_projet.services.CardService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,30 +11,28 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.event.ApplicationStartedEvent;
 import org.springframework.context.event.EventListener;
 
-import java.sql.Timestamp;
-import java.time.Instant;
 import java.util.List;
 
-import com.example.gwent_projet.services.UserService;
-import com.example.gwent_projet.utils.consoleDisplay;
 
 @SpringBootApplication
 public class GwentProjetApplication {
 
     @Autowired
     private CardService cardService;
+    @Autowired
+   private UserCommand userCommand;
 
     public static void main(String[] args) {
         SpringApplication.run(GwentProjetApplication.class, args);
         
-		UserService userService = new UserService();
+		
 		
 		// ------------------------------------------------------------------
 		/* temporary console display for user handling
 		 * feel free to delete once integrated into framework 
 		 */
 		// execute method "run" that prints the menu and handles actions per user choice
-		userService.run();
+		
 		// line break because Java sucks at handling line breaks
 		
     }
@@ -44,5 +43,6 @@ public class GwentProjetApplication {
         System.out.println("GwentApp is started !");
         System.out.println("--------------------");
         List<Card> cards = cardService.findAll();
+        this.userCommand.run();
     }
 }
