+++ src/main/java/com/example/gwent_projet/dto/CreateUser.java
@@ -1,0 +1,11 @@
+package com.example.gwent_projet.dto;
+
+// ORM
+public class CreateUser {
+
+	public int role;
+	public String username;
+	public String email;
+	public String password;
+	
+}
